+++ packages/components-js/tests/e2e/specs/flyout.e2e.ts
@@ -797,7 +797,7 @@ test.describe('after dynamic slot change', () => {
 
 test.describe('events', () => {
   skipInBrowsers(['firefox', 'webkit']);
-  test('should expose ontransitionstart event', async ({ page }) => {
+  test('should call motionVisibleEnd event when opening transition is finished', async ({ page }) => {
     await initBasicFlyout(
       page,
       { open: false },
@@ -807,16 +807,15 @@ test.describe('events', () => {
     );
     const host = await getHost(page);
     await waitForStencilLifecycle(page);
-    await addEventListener(host, 'transition');
+    await addEventListener(host, 'motionVisibleEnd');
 
-    expect((await getEventSummary(host, 'transition')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(0);
 
     await openFlyout(page);
 
-    // Transition for each property will be fired separately (overlay, box-shadow, opacity, backdrop-filter, background-color, transform, margin-right)
-    expect((await getEventSummary(host, 'transition')).counter).toBe(7);
+    expect((await getEventSummary(host, 'motionVisibleEnd')).counter).toBe(1);
   });
-  test('should expose ontransitionend event', async ({ page }) => {
+  test('should call motionHiddenEnd event when closing transition is finished', async ({ page }) => {
     await initBasicFlyout(
       page,
       { open: true },
@@ -826,13 +825,12 @@ test.describe('events', () => {
     );
     const host = await getHost(page);
     await waitForStencilLifecycle(page);
-    await addEventListener(host, 'transition');
+    await addEventListener(host, 'motionHiddenEnd');
 
-    expect((await getEventSummary(host, 'transition')).counter).toBe(0);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(0);
 
     await dismissFlyout(page);
 
-    // Transition for each property will be fired separately (overlay, box-shadow, opacity, backdrop-filter, background-color)
-    expect((await getEventSummary(host, 'transition')).counter).toBe(5);
+    expect((await getEventSummary(host, 'motionHiddenEnd')).counter).toBe(1);
   });
 });