+++ packages/components/src/components/modal/modal.tsx
@@ -22,12 +22,17 @@ import {
   warnIfDeprecatedPropIsUsed,
 } from '../../utils';
 import type { ModalAriaAttribute, ModalBackdrop } from './modal-utils';
-import { MODAL_ARIA_ATTRIBUTES, type ModalTransitionEventDetail } from './modal-utils';
+import {
+  MODAL_ARIA_ATTRIBUTES,
+  type ModalMotionVisibleEndEventDetail,
+  type ModalMotionHiddenEndEventDetail,
+} from './modal-utils';
 import { getComponentCss } from './modal-styles';
 import { BACKDROPS } from '../../styles/dialog-styles';
 import { getSlottedAnchorStyles } from '../../styles';
 import { observeStickyArea } from '../../utils/dialog/observer';
 import { getDeprecatedPropOrSlotWarningMessage } from '../../utils/log/helper';
+import { onTransitionEnd } from '../../utils/dialog/dialog';
 
 const propTypes: PropTypes<typeof Modal> = {
   open: AllowedTypes.boolean,
@@ -95,8 +100,11 @@ export class Modal {
   /** Emitted when the component requests to be dismissed. */
   @Event({ bubbles: false }) public dismiss?: EventEmitter<void>;
 
-  /** Emitted when the native ontransitionstart or ontransitionend event is fired from the dialog element. */
-  @Event({ bubbles: false }) public transition?: EventEmitter<ModalTransitionEventDetail>;
+  /** Emitted when the modal is opened and the transition is finished. */
+  @Event({ bubbles: false }) public motionVisibleEnd?: EventEmitter<ModalMotionVisibleEndEventDetail>;
+
+  /** Emitted when the modal is closed and the transition is finished. */
+  @Event({ bubbles: false }) public motionHiddenEnd?: EventEmitter<ModalMotionHiddenEndEventDetail>;
 
   private dialog: HTMLDialogElement;
   private scroller: HTMLDivElement;
@@ -197,8 +205,7 @@ export class Modal {
         onCancel={(e) => onCancelDialog(e, this.dismissDialog, !this.hasDismissButton)}
         // Previously done with onMouseDown to change the click behavior (not closing when pressing mousedown on modal and mouseup on backdrop) but changed back to native behavior
         onClick={(e) => onClickDialog(e, this.dismissDialog, this.disableBackdropClick)}
-        onTransitionStart={(e) => this.transition.emit(e)}
-        onTransitionEnd={(e) => this.transition.emit(e)}
+        onTransitionEnd={(e) => onTransitionEnd(e, this.open, this.motionVisibleEnd, this.motionHiddenEnd)}
         {...parseAndGetAriaAttributes({
           'aria-modal': true,
           'aria-label': this.heading,